src/main/java/org/cyk/system/poulsscolaire/server/api/accounting/FundingSourceFilter.java
package org.cyk.system.poulsscolaire.server.api.accounting;

import ci.gouv.dgbf.extension.server.service.api.AbstractIdentifiableFilter;
import ci.gouv.dgbf.extension.server.service.api.request.FilterDto;
import lombok.Getter;
import lombok.Setter;

/**
 * Cette classe représente le filtre de {@link FundingSourceDto}.
 *
 * @author Christian
 *
 */
@Getter
@Setter
public class FundingSourceFilter extends AbstractIdentifiableFilter {

  String schoolIdentifier;

  public FundingSourceFilter(FilterDto dto) {
    super(dto);
  }

  public FundingSourceFilter() {}

  @Override
  protected void doInitialize(FilterDto filter) {
    super.doInitialize(filter);
    schoolIdentifier = getSchoolIdentifier(filter);
  }

  @Override
  protected void toDto(FilterDto filter) {
    super.toDto(filter);
    setSchoolIdentifier(filter, schoolIdentifier);
  }

  public static void setSchoolIdentifier(FilterDto filter, String identifier) {
    set(filter, JSON_SCHOOL_IDENTIFIER, f -> f.getValueAsString(),
        f -> f.setValueAsString(identifier));
  }

  public static String getSchoolIdentifier(FilterDto filter) {
    return get(filter, d -> d.getFieldValueAsStringByName(JSON_SCHOOL_IDENTIFIER));
  }

  public static final String JSON_SCHOOL_IDENTIFIER = FundingSourceDto.JSON_SCHOOL_IDENTIFIER;  
}
